Move logic to 'CTypeFromDeclaration' from 'StringConstruction'.
'StringConstruction' is a bad name; but also, the class was doing the
work of both an AST Listener, as well as having this logic.

Another result of this was that classes like Insturmentor or
TypeInference, which happen to need Scopes also, 'hid' this dependency
by going through StringConstruction.

'StringConstruction' is still a bad name for the phase/pass of the AST,
though.

+++ src/test/scala/edu/nus/worksheet/instrumentor/test/TypeInferenceSpec.scala
@@ -165,7 +165,7 @@ int* f1() {
     assertInference(PrimitiveType("!i", "int"), "int i;", "!i");
     assertInference(PointerType("&i", PrimitiveType("i", "int")), "int i;", "&i");
     assertInference(PrimitiveType("(*p)", "int"), "int *p;", "*p");
-    assertInference(PrimitiveType("sizeof p", "size_t"), "int p;", "sizeof p");
+    assertInference(PrimitiveType("sizeof p", "int"), "int p;", "sizeof p");
   }
 
   it should "infer cast expressions" in {
